crown_rad(likelihood)	R Documentation

Dataset of Tree DBH and Crown Radius

Description

This is a set of imaginary data for DBH and crown radius for a set of trees.

Usage

crown_rad

Format

Tab-delimited text.

[Package likelihood version 1.3 Index]