likelihood-package(likelihood)	R Documentation

Package for maximum likelihood estimation

Description

This package allows you to use R to write complex scientific models, and then estimate maximum likelihood for parameters using simulated annealing.

Details

Package: 	 likelihood

Version: 	 1.2

Date: 	 2006-12-15

Depends: 	 R (>= 2.1.1, nlme

License: 	 GNU Public License, see the file COPYRIGHT for details

Several demonstration scripts are included in the demo directory.

Author(s)

Lora Murphy <murphyl@ecostudies.org>

Maintainer: Lora Murphy <murphyl@ecostudies.org>

See Also

neighlikeli for a package of additional functions for maximum likelihood estimation of neighborhood models in forest ecology.

[Package likelihood version 1.1 Index]