Previous Users(likelihood)	R Documentation

Changes for Users of likelihood package 0.1

Description

Changes that must be made for scripts prepared under likelihood package 0.1

Critical Change

R scripts that use the anneal function will mostly not work in likelihood 1.0 without one change. The par list argument for anneal and likeli has been split into two list arguments, par and var. par is for model parameters, those values that will be varied in a simulated annealing run. var is for all other non-varying data and arguments needed by the model. Since this distinction is unimportant when using the function likeli alone, scripts using only that function will work as-is, although you should still update them if possible.

To fix a script for anneal, note the list elements for the par_lo and par_hi arguments. These are the elements that remain in par. Place all other elements in par in a new list, and pass it as var.

Other New Options

anneal can calculate the parameter variance / covariance matrix and standard errors, and add these to its output. Use the new anneal argument hessian = TRUE.

anneal can perform a global search on parameters. Omitting a bound means that the search occurs to positive or negative infinity, or both.

You have more control over how anneal calculates support limits, with the new arguments delta and slimit.

[Package likelihood version 1.2 Index]