write_results(likelihood)	R Documentation

Write the Results of Simulated Annealing to File

Description

Takes the results produced by the function anneal and writes them to a tab-delimited text file.

Usage

write_results(results, filename, data = TRUE, print_whole_hist = FALSE)

Arguments

results 	 The output list produced by the function anneal.

filename 	 A string with the file and path to the file you wish to write. This will overwrite any existing files of that name. This will not add any file extensions so remember to put on the appropriate one.

data 	 If TRUE, the source_dataset member of results is written to the file; if FALSE, it is not. Large datasets can inflate the size of output files.

print_whole_hist 	 If TRUE, the entire likelihood history of the run is printed; if FALSE, it is not. Long runs can have rather long histories.

Value

A file with the contents of results written as tab-delimited text.

See Also

anneal

Examples

## Assuming you have performed a simulated annealing run and placed the 
## results in an object called "my_results"...
## Not run: write_results(my_results, "c:\results.txt")

[Package likelihood version 1.3 Index]